+++ pelutils/datahandling.py
@@ -1,17 +1,33 @@
 from __future__ import annotations
+from collections import defaultdict
 import os
 import json
+import pickle
 
 import numpy as np
+try:
+    import torch
+    _has_torch = True
+except ModuleNotFoundError:
+    _has_torch = False
 
+# The special serializations that DataStorage supports.
+# datatype: tuple consisting of save function (API is save(file, data)), load function (API is load(file_obj)) and extension
+SERIALIZATIONS = {
+    np.ndarray: (np.save, np.load, "npy"),
+}
+if _has_torch:
+    SERIALIZATIONS[torch.Tensor] = (lambda f, d: torch.save(d, f), torch.load, "pt")
 
 class DataStorage:
     """
     The DataStorage class is an augmentation of the dataclass that incluces a standard way to save and load data.
 
-    Currently works with:
+    Currently works specifically with:
         * Numpy arrays (numpy.ndarray)
+        * Torch tensors (torch.Tensor)
         * Any json serializable type - that is, it should be savable by json.dump
+    * All other data structures are pickled
 
     The field `subfolder` is the directory in which to place all saved date
     `json_name` chooses the name of the single json data file including all jsonifiable data
@@ -37,11 +53,13 @@ class DataStorage:
     ```
     """
 
-    subfolder = ''
-    json_name = 'data.json'
+    subfolder = ""
+    json_name = "data.json"
+    pickle_ext = "p"
 
     def save(self, loc: str = '') -> list[str]:
-        """Saves all the fields of the instatiated data classes in npy's or json in <`loc`>/<`self.subfolder`>
+        """
+        Saves all the fields of the instatiated data classes as either json, pickle or designated serialization function
 
         :param str loc: Save location to place the subfolder in
         :return: list (with length = # of saved files) of full save paths
@@ -50,22 +68,36 @@ class DataStorage:
         if loc:
             os.makedirs(loc, exist_ok=True)
 
-        # Split data by whether it should be saved .json and to .npy
-        to_json, to_npy = dict(), dict()
+        # Split data by whether it should be saved using a known function or using pickle or json
+        func_serialize = defaultdict(dict)
+        to_pickle, to_json = dict(), dict()
         for key, data in self.__dict__.items():
-            if isinstance(data, np.ndarray):
-                to_npy.update({key: data})
+            for datatype in SERIALIZATIONS:
+                if isinstance(data, datatype):
+                    func_serialize[datatype][key] = data
+                    break
             else:
-                to_json.update({key: data})
-
+                try:
+                    # Test whether the data is json serializable by dumping it to string.
+                    json.dumps({key: data})
+                    to_json[key] = data
+                except TypeError:
+                    to_pickle[key] = data
         # Save data
-        paths = [os.path.join(loc, self.json_name)]
+        paths = list()
         if to_json:
-            with open(paths[0], "w", encoding="utf-8") as f:
-                json.dump(to_json, f, indent=4)
-        for key, arr in to_npy.items():
-            paths.append(os.path.join(loc, f"{key}.npy"))
-            np.save(paths[-1], arr)
+            paths.append(os.path.join(loc, self.json_name))
+            with open(paths[-1], "w", encoding="utf-8") as f:
+                json.dump(to_json, f)
+        for key, data in to_pickle.items():
+            paths.append(os.path.join(loc, f"{key}.{self.pickle_ext}"))
+            with open(paths[-1], "wb") as f:
+                pickle.dump(data, f)
+        for seri, datas in func_serialize.items():
+            save, _, ext = SERIALIZATIONS[seri]
+            for key, data in datas.items():
+                paths.append(os.path.join(loc, f"{key}.{ext}"))
+                save(paths[-1], data)
         return paths
 
     @classmethod
@@ -76,21 +108,28 @@ class DataStorage:
         :return: An instance of this class with the content of the files
         """
         loc = cls._get_loc(loc)
-        any_json = False
-        npys, non_numpy = dict(), dict()
-
-        # Get .npy filenames and load these
+        fields = dict()
+        # List of fields non-loadable using the SERIALIZATIONS functions
+        generals = list()
         for key, field in cls.__dict__["__dataclass_fields__"].items():
-            if field.type == np.ndarray:
-                npys[key] = np.load(os.path.join(loc, f"{key}.npy"))
+            for datatype in SERIALIZATIONS:
+                if field.type  == datatype:
+                    _, load, ext = SERIALIZATIONS[datatype]
+                    fields[key] = load(os.path.join(loc, f"{key}.{ext}"))
+                    break
+            else: generals.append(key)
+        # Check if the field was saved with pickle
+        any_json = False
+        for key in generals:
+            pfile = os.path.join(loc, f"{key}.{cls.pickle_ext}")
+            if os.path.isfile(pfile):
+                with open(pfile, "rb") as f:
+                    fields[key] = pickle.load(f)
             else: any_json = True
-
-        # Load .json if any non ndarray fields exist
         if any_json:
             with open(os.path.join(loc, cls.json_name), encoding="utf-8") as f:
-                non_numpy = json.load(f)
-
-        return cls(**non_numpy, **npys)
+                fields.update(json.load(f))
+        return cls(**fields)
 
     @classmethod
     def _get_loc(cls, loc: str):