+++ HISTORY.md
@@ -5,6 +5,7 @@
 - Test logging now uses `Levels.DEBUG` by default
 - Added `TickTock.fuse_multiple` for combining several `TickTock` instances
 - Fixed bugs when using multiple `TickTock` instances
+- Allow multiple hits in single profile
 - Now possible to profile using `with` statement
 - Added method to logger to parse boolean user input
 - Added method to `Table` for adding vertical lines manually